requirements.txt
# FastMCP framework
fastmcp>=2.0.0

# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
asyncio-throttle>=1.0.2

# External API clients
openai>=1.0.0
pinecone>=5.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# Utilities
markdown>=3.4.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
